Remove exampleapp
