+++ Makefile
@@ -22,20 +22,6 @@ upgrade: ## update the *requirements.txt files with the latest packages satisfyi
 	sed 's/==/>=/g' requirements.txt > requirements.tmp
 	mv requirements.tmp requirements.txt
 
-clean_examples:
-	# Remove the tables specific for the example app
-	python backend/manage.py migrate exampleapp zero
-	# Removing backend example app files
-	rm -rf ./backend/exampleapp
-	# Removing frontend example app files
-	rm -rf ./frontend/js/app/example-app
-	# Removing example templates files
-	rm -rf ./backend/templates/exampleapp
-	# Remove exampleapp from settings
-	sed -i '/exampleapp/d' ./backend/{{project_name}}/settings/base.py
-	# Remove exampleapp from urls
-	sed -i '/exampleapp/d' ./backend/{{project_name}}/urls.py
-
 compile_install_requirements:
 	@echo 'Installing pip-tools...'
 	export PIP_REQUIRE_VIRTUALENV=true; \
