+++ backend/project_name/settings/base.py
@@ -25,7 +25,6 @@ AUTH_USER_MODEL = "users.User"
 ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
-    "exampleapp.apps.ExampleappConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
