+++ backend/templates/common/index.html
@@ -1,0 +1,13 @@
+{% extends 'base.html' %}
+
+
+{% block body %}
+  <h1>Django React Boilerplate Example App</h1>
+
+  <p>The setup of the Django React Boilerplate Example App was successful.</p>
+
+  <h2>React Example App</h2>
+  <p>Here is an example of a React app included in the Django framework:</p>
+
+  <div id="react-app"></div>
+{% endblock %}
