--- backend/templates/exampleapp/results.html
@@ -1,9 +1,0 @@
-<h1>{{ question.question_text }}</h1>
-
-<ul>
-  {% for choice in question.choice_set.all %}
-    <li>{{ choice.choice_text }} -- {{ choice.votes }} vote{{ choice.votes|pluralize }}</li>
-  {% endfor %}
-</ul>
-
-<a href="{% url 'exampleapp:detail' question.id %}">Vote again?</a>
